+++ vm/native/java/util/concurrent/atomic/atomic_long.go
@@ -3,6 +3,6 @@ package atomic
 import "github.com/murakmii/gj/vm"
 
 func AtomicLongVMSupportsCS8(thread *vm.Thread, _ []interface{}) error {
-	thread.CurrentFrame().PushOperand(0)
+	thread.CurrentFrame().PushOperand(int32(0))
 	return nil
 }
